final v0.0.1

+++ dotNetEventManagement/Controllers/OrderController.cs
@@ -120,7 +120,7 @@ namespace dotNetEventManagement.Controllers
                             FullName = reader.GetString(reader.GetOrdinal("FullName")),
                             EventId = reader.GetString(reader.GetOrdinal("EventId")),
                             EventName = reader.GetString(reader.GetOrdinal("EventName")),
-                            TotalPrice = (reader.GetDecimal(reader.GetOrdinal("TotalPrice"))),
+                            TotalPrice = (double)(reader.GetDecimal(reader.GetOrdinal("TotalPrice"))),
                             OrderDate = reader.GetDateTime(reader.GetOrdinal("OrderDate")),
                             PaymentStatus = reader.GetString(reader.GetOrdinal("PaymentStatus"))
                         });
@@ -151,7 +151,7 @@ namespace dotNetEventManagement.Controllers
                             FullName = reader.GetString("FullName"), 
                             EventId = reader.GetString("EventId"),
                             EventName = reader.GetString("EventName"),
-                            TotalPrice = (reader.GetDecimal(reader.GetOrdinal("TotalPrice"))),
+                            TotalPrice = (double)(reader.GetDecimal(reader.GetOrdinal("TotalPrice"))),
                             OrderDate = reader.GetDateTime("OrderDate"),
                             PaymentStatus = reader.GetString("PaymentStatus")
                         });
@@ -256,7 +256,7 @@ namespace dotNetEventManagement.Controllers
                             FullName = reader.GetString(reader.GetOrdinal("FullName")),
                             EventId = reader.GetString(reader.GetOrdinal("EventId")),
                             EventName = reader.GetString(reader.GetOrdinal("EventName")),
-                            TotalPrice = (Decimal)reader.GetDecimal(reader.GetOrdinal("TotalPrice")),
+                            TotalPrice = (double)reader.GetDecimal(reader.GetOrdinal("TotalPrice")),
                             OrderDate = reader.GetDateTime(reader.GetOrdinal("OrderDate")),
                             PaymentStatus = reader.GetString(reader.GetOrdinal("PaymentStatus"))
                         };
